refactor: replace chat service with api calls
